moved public board to its own folder
